Organize previous code into packages
Previously, the bulk of the code resided in the main Duke driver class,
and OOP was only employed to model the Calendar and Tasks (Deadline,
Todo, Event).

To better to adhere to OOP principles for this software engineering
project. The code was refactored into different packages. All the code
is in the overarching duke package, and each subpackage aims to serve a
specific function for the program. E.g. duke.info stores all the
information required for the program (Calendar, Tasks, Exceptions) and
the duke.ui package handles the reading and writing from user input.

Let's
* Refactor the code into separate packages for better organization
* Use parent abstract classes to abstract out common functionalities
* Use a Text file to store all standard output messages