Fix bug in "unmark" command
Previously, there was a bug where the unmark command did not mark the
job as incomplete, but instead marked the job as complete due to an
erroneous call of markAsDone() instead of markAsNotDone().

This commit changes the method call to properly reflect the action of
marking the task as incomplete.

Let's
* Fix the unmark command to properly carry out the desired functionality