Add save functionality when Calendar chanes
Previously, the list of tasks in the Calendar was deleted when the
program exits. This meant that the user would lose all the information
about the tasks after each run of the program.

To make the usage of the program more meaningful, a save functionality
was added that saves the current list of tasks to "tasks.txt" in the
main program folder whenever the task list is updated. This makes the
program more meaningful to use, as there is persistent information that
users can use between each program run.

Let's
* Add a save function to allow users to keep their progress
* Add a load function to automatically load a previous save on startup
* Use a "tasks.txt" file in the main directory to store progress