Add "Save/Load" of list data
Currently the Arraylist of Tasks in Duke.java is lost after the program exits.

This makes the program vulnerable to power loss and accidental exits.

Therefore this commit updates Duke.java to save the Arraylist data to "data/data.bin" and load from it if the file exists on start.
It also updates Task.java to make the class implement Serailizable for this purpose.