Add packages to files in project
This is to hop on this actual/real commit and to test writing in the body of a commit to see if everything works out fine.
This commit divide classes in Duke project into packages.
Organising the classes into suitable java packages
Ah I missed a fullstop... I can't revert ugh git CLI.
Minimal and stretch goals completed!
Alright let's see if it works on GitHub.