Add Command class
Commands entered by User is parsed by main Duke class.

Extracting the various commands into a seperate Command class
will reduce spagetti code and provides the use of inheritance
for various command types.

Let's:
* Create A new class that represents the various actions that
are available to the user.

This will allow the use of polymorphism when executing commands.

* Note: only skeletal code is added for now, implementations
will follow in future commits.
* Parser for commands is also added.
* Main Duke class does not use Command class yet, will refactor
it to use Commands class once it is functional.
* Minor changes in Storage and Ui are also included in this commit.