Patch on level 7
Previously, save command could only save for viewing in text file.
To allow for automatic saving and loading, text storage and text
loader was created.

Added a new InvalidOperationException for when program tries to
run mark on an existing marked task, and unmark on an unmarked
task.

Refactored exceptions such that most exceptions caught in execute
commands are caught in Command Manager instead.