Duke class: Update Duke's response to "invalid keywords"
Duke's response to "invalid keywords" is to print a list of
recognized keywords that the user can use.

Updating the response will teach users the keywords they can use to
access the latest features of Duke.

Let's update the printed response to correctly reflect the current
list of recognized keywords.