Commands class: add startup() method
There does not exist a way for us to detect if ./data/duke.txt is a
valid file path that exists.

In order for us to implement our "save" feature in Duke, we require
./data/duke.txt to exist so we can update the duke.txt file after
each update by the user.

Let's implement a startup() method that is called at the beginning
of each session with Duke that checks for and creates the data
folder and duke.txt file.

This is so we can easily bug fix this startup() method in the future
in the case of any poor implementation.

The startup() method currently implemented is only able to check for
the existence of a folder named "data" in the root folder containing
Duke.java.