Commands class: update startup() method
startup() method is only capable of checking the existence of a
"data" directory in the root directory containing the Duke program.

This is only a means to determine if there is a need to manually
create the directory for future use.

Let's now provide an independent path to contain our duke.txt file
and also create the parent directory, child directory and duke.txt
file if it does not already exist.

startup() now creates a folder and duke.txt file that will be used
to keep track of a users task history.

The specific path used is "C:\Users\username\Duke\data". The main
method in Duke.java has also been updated to call startup() at the
beginning of each session.