Parser class: Add new Parser class
Require a Parser class to deal with making sense of the user command.

Let's:
* Add a Parser class to parse user commands, when the command needs
to be split into multiple parts.

Additional Remarks: Did not quite get what a parser class is useful
for, as most parts of the code are rather blended together and cannot
be shifted to Parser. May require future changes to perfect Parser.