Commands class: add restore() method
There does not exist a way for users to reuse a previously created
local copy of duke.txt.

This does not fulfil one of the features of Duke that we are trying
to implement. We want to be able to load the data from the hard disk
when Duke starts up

Let's now provide a way to load the contents of an existing duke.txt
file onto a session of Duke.

By creating a restore() method that takes in a File object as
argument, we are able to customize how Duke handles a custom
duke.txt script. This gives us the flexibility to format the
duke.txt file for our convenience.

To facilitate the implementation of restore(), we want to have
special constructors for the Todo, Deadline and Event classes that
can allow us to create tasks based on the information stored on the
duke.txt file.

Lastly, we also want to provide methods to add these tasks on the
duke.txt file in the background without interfering with the current
session of duke. This is done by overloading the add"" methods in
the Commands class