Extract file loading behavior to a seperate interface
This helps to make Storage.java more generic. Allows other class type to be able to use storage.java to store data, and also makes the class responsible for how it extracts data from the file for itself. In Duke.java made a lambda expression to pass to storage so it can create the proper task depending on what data received.