Commands class: add update() method
There does not exist a way for users to manually request Duke to
store the data of the current session onto their local device.

We want to provide the flexibility to our users to manually request
Duke to store the data of their current session. We also need a way
to trigger this storing of data when Duke is closed so this is a
step towards that.

Let's now provide a way to request Duke to update the local duke.txt
file with the latest records stored in Duke for the current session.

By creating a update() method, we are able to call for the update in
our code and also provide a way for the user to manually request for
this update without having to close Duke.

The ToDos, Deadline and Event classes have been provided methods to
help with update(). Also, a new keyword "update" is recognized by
Duke.