Duke class: update "bye" keyword behavior
Previously, we do not get Duke to update the local duke.txt file at
the end of a session.

If the user forgets to call "update" manually, the data created
or changed during the session is not saved.

Let's make it so that Duke will automatically save data to the user
hard disk by calling update() when "bye" keyword is used.