Created Checker class
Checker class now handles the enum, allowing for switch case use in Duke
to keep the code as tidy as possible. All other functions, such as the
addition of deadlines or todos, still remain the same.

Added a new string to set as the String to be returned when no
description is to be found. Provides ease of change for the future.

JavaDocs needs to be done for the Checker class.