Duke class: add isDateValid() method
For keywords "deadline" and "event", Duke interprets the date & time
arguments provided as Strings and stores them word for word in it's
records.

We want to provide users with a more uniform response when it comes
to date & time such that any provided input will give the same
output.

Let's first introduce a method to validate date formats so we can
teach Duke different date formats and get it to recognize them.

We want to create a method that we can expand on in the future,
adding new date formats for Duke to recognize.

Currently, the method only has 2 date formats that it recognizes and
does not validate time.