Duke class: change isDateValid() method to convertToDukeDate() method
isDateValid() returned a boolean that indicated if Duke recognized
the given date format or not.

This boolean needed to be handled separately outside of
isDateValid() before Duke will process the given date into a uniform
format.

Let's change the isDateValid() method into a method that immediately
does the conversion as well and returns a LocalDate object.

This gives as a more useful object to manipulate accordingly in our
other methods in the Commands class.

convertToDukeDate() has been moved from Duke.java to Commands.java
where it will be better used. Both "deadline" and "event" keywords
have had their implementations updated to include the use of
convertToDukeDate().