Duke class: add convertToDukeTime()
Duke currently treats time as a String and does not manipulate it in
any way.

We want Duke to be able to convert time inputs in 24-hour format to
a standardized output.

Let's implement a new method convertToDukeTime() that does this.

This gives us the flexibility to detect and convert 24-hour time
to 12-hour time for any commands that expect a time input from the
user.

"deadline" and "event" keywords have been updated to call
convertToDukeTime() when needed.