Parser Class: Move some codes in main Duke class to Parser class
Some code that are more suitable for Parser class to handle are in
the main Duke class.

This can potentially cause dependability, as the main class is handling
some of the Parser class's codes.

Let's:
* Move the relevant parts to Parser class, mainly the exception
checking and trimming of results
* Ensure that the result passed back to the main Duke class is
processed and ready for use, which corresponds to the purpose of a
Parser class.
* Add a new exception IncompleteArgumentException for Parser exceptions.