Added date functionality to Deadline and Event
Created a new class "DateHelper" that helps with converting strings to
Date class. DateHelper stores the dates and can format the date into
the required formats to be outputted in Deadline and Event.

Updated input.txt and EXPECTED.TXT to include the new date formats.

JavaDocs to be updated for DateHelper, TextWriter, and Duke.