Support building as JAR file.
The current method of accessing resource files only worked when running
the program from IntelliJ.
The program was not able to access resource files when ran as a JAR.
This caused the program to not be able to load, save or play audio.

Wrap all file streams with a buffered stream.

This lets the program to access resource files regardless of how it is
packaged.