Inconsistent "From Duke: " present in some commands, while absent from some Lack Text UI testing General print statements for caught errors
Need to change to have consistent in the way Duke responds to commands.
To allow Testing of the program
More specific to allow users to know why their inputted command is wrong

Right after the scanner scans for next line, Have a print statement. "From Duke"
Changed the input.txt file as well as the EXPECT.TXT file to ensure that ACTUAL.TXT will be equal to EXPECTED.TXT when input.txt in fed into the bash