Duke class: update Duke to save state after every command
Duke currently only saves the state to the user's local hard drive
on request ("update") or at the end of a session.

This could led to potential lost of user data if an expected crash
leading to an unexpected end to the session without the user or
Duke invoking the update() command.

Let's make it so that Duke will automatically try to update the
local duke.txt file after every command.

We are simply adding a update() call after every keyword call.