Refactor list of tasks into separate class
Previously, an `ArrayList<Task>` was used to contain the list of tasks.
Methods were then declared to manipulate the items in `ArrayList`

However, this is not good practice of encapsulation.

Thus, behaviour involving manipulating the list of tasks and
the `ArrayList<Task>` instance have been refactored into their separate
class.