Completed OOP of Parser, Storage, TaskList, and Storage
Edited all files to help keep Duke as lean as possible. Parser is an
enum class that helps to check which command was given. Storage helps
read and write files. Every change in the list will be marked down in
the textfile immediately. Time is also handled in a DateHelper.

JavaDocs to be updated and do error handling for mark and unmark. Mark
and unamrks' error in id passed is not properly handled.