Add Date functionality to Duke
Currently, dates are stored merely as strings in each Task. This means
that the dates cannot be manipulated in a meaningful way and prevents
any further improvement to the date functionality (such as filtering
tasks by date)

To make the date formatting more useful, the date attribute now utilizes
the LocalDate library to parse dates. When a date is input, the Task()
constructor will attempt to parse it into a LocalDate format. If the
date cannot be parsed, it will be treated as a simple string.

Lets
* Use the LocalDate library to allow dates to be utilized more
  functionally