Add Find Functionality to Duke
Currently, the list of tasks can only be shown statically, which means
that as the list of tasks scale up and there are a large amount of
tasks, it would be difficult for a user to search for a specific task or
any tasks matching a certain criteria.

To make the calendar more useful, FindCommand is added and the user can
input "find <keyword>" to list all tasks that match that particular
keyword.

Lets
* Add a find function to allow users to find tasks matching a keyword