Add Command class and its subclasses
Currently all the functionality of the program is handled internally in
Duke.java

This  increases the difficultly in adding and editing commands while
maintaining code correctness.

Therefore, this commit will add a new Command class as well as sub
classes for the various commands to make the project more
Object oriented.