Add JavaDocs documentation to code
Currently, there is no documentation for the codebase. This makes it
difficult for other developers to contribute to the project or to fix
bugs related to the codebase in the future.

To make the codebase more informative, JavaDocs documentation has been
added to describe each method and ease future development of the
codebase

Lets
* Add JavaDocs documentation to all methods in the codebase