Add JUnit tests for Todo
Currently, there are no unit tests for the classes in the package. This
means that there is no confirmation that any of the classes is working
as intended, which might result in regression bugs in the future when
the code is extended.

To test the codebase, JUnit tests for Todo was added to check that the
methods in the Todo class are working as intended.

Lets
* Add preliminary JUNit tests for the Todo class