Added Command class and Packages
Added Command class which is an abstract class. This class is made to
handle the different commands that could be executed by Ducky. Classes
include:
- AddEventCommand
- AddDeadlineCommand
- AddTodoCommand
- MarkCommand
- UnmarkCommand
- ExitCommand
- ListTaskCommand
- HelpCommand
- DeleteCommand

Packages were also created to add in classes of the same types. Three
packages created were:
- Duke
- Command
- Task
Command and Task are packages under the Duke package.

JavaDocs needs to be completed further.