Duke package: add support for "find" keyword.
The only way for users to view the records in Duke during a session
is to use the "list" keyword.

This is inefficient and provides very little flexibility when it
comes to viewing records.

Let's add a feature that allows the user to view all tasks in Duke's
records according to a specified word or phrase.

We will implement a new keyword "find" that takes in a phrase, looks
through duke's history and returns a list of all tasks that has that
phrase in their description.

Support for "find" keyword has been added in subclasses supporting
the Duke package in the form of:
Commands class: add find() method
DukeHistory class: add findPhrase() method
DukeUi class: add printFoundNothing() method