Extract Ui, Storage, Parser and TaskList classes
The current driver in the main method lumps everything that needs to
be done, there is no good OOP practice in place.

Let's create separate classes to encapsulate the different methods that
need to be called.

Separating the different methods into the relevant classes help to make
the code more readable and will help with scalability.