Fixed Minor Bugs
Adjusted access modifier for methods that are only access by the class
itself.

Updated parser to replace all "|" instance to "-" to ensure it does not
affect saving to local persistent storage.

Added missing constants and methods for Find implementation due to
merge.

Fixed Index Out of Bounds for Done command due to wrong order of if
statement to check out of range item.