Added Gradle to automate project builds
Gradle is now set up and can build and run Duke. To run it, go to
Tasks > application > right click 'run' and run it. This will start the
Ducky CLI.

Gradle can also run JUnit tests. There was an error in the
`./gradlew clean test` initially, where a class could not be
initialized. Error was resolved by changing the Gradle version to 6.4.1.