Add GUI for Duke
Duke only supports a text based UI.

This makes the program unappealing and dull to use as it is hard
to convey character through text alone.

A GUI helps make the program more appealing to use and allows for
creative design choices in the future.

Let's do the following to add a GUI for Duke using JavaFX,
* Modify the Command classes to allow them to modify StringBuilder
objects.
* Move Duke.java into the duke package.
* Change the return type of Ui methods from void to String to pass
to the new GUI classes to display.
* Modify UiTest.java to reflect the changes made in Ui.java.
* Add new GUI classes: Main, Launcher, DialogBox, MainWindow.