Use assert to document important assumptions
An assertion failure indicates a possible bug in the code because
the code has resulted in a program state that violates an assumption
about how the code should behave. An assertion failure indicates a
possible bug and the sooner the execution stops, the safer it is.

Assertions are used in following places:
* check the string return from checkEmptyTask() to make sure it's not empty
* check the index return from parseListIndex() to make sure it's not out of range
* check format return from parseTaskFormat() to mark sure the format is correct