A-Assertions
Add assertions at certain points of the code.

Assertions are used to ensure correctness of the code at certain points

One of the few important points are creation of new Commands. Ensure that new Commands are of the correct, specific type (not just a Command).

As a step to ensure correctness in the code base, let's add assertions at these important points. Doing so will ensure that code correctness is maintained as much as possible before new commits.