C-DetectDuplicates
Add a DuplicateCommand class and logic to identify when a duplicate Task is about to be added.

Logically, duplicate tasks should not be added to the list. Therefore, it is necessary to handle the case where a duplicate item is detected.

A duplicate item will not be added to the TaskList. Instead, a message (GUI) will be shown to the user to let them know they are adding a duplicate task.

This can be even more comprehensive by detecting if there are conflicting date/times. Doing so will ensure the user is more aware of their schedule.