Assertions: Calling Assertions to make checks to assumptions
Some codes require us to make certain assumptions for it to work as intended.

Creating Assertions to make sure that all these assumptions are always true reduces the amount of bugs that can pop up.

Let's create some Assertions in relevant classes to reduce failure of assumptions.

Use junit asserttrue to check if a particular process is true.

Refer to this website for additional information on how to use assertTrue: https://www.baeldung.com/java-file-directory-exists