DialogBox constructors: add assert statements
Constructor in DialogBox does not check that the input parameters are not null.

When null values are passed into DialogBox constructors, an unintelligible error message is printed to the console, making it difficult for developers to debug.

This situation may occur when the image to be included in the dialog box fails to be loaded by the program. This violates the basic assumption that all the required assets will be successfully loaded by the program.

To help make such an error easier to spot, let's add an explicit assertion to check that the image object is non-null.