Implement assertions
Currently no assertions are used to ensure certain assumptions hold.

Need to include the assertion so that the code will not crash when
certain assumptions are missed, and inform the author of the code, which
assumptions are violated.

Let's include assertion in different parts of the code:
- AddCommand.java : assert to ensure exceptions from adding tasks are
  caught
- DeleteCommand.java : assert to ensure delete exceptions are caught
- FindCommand.java : assert to ensure find details are not empty
- MarkCommand.java : assert to ensure exceptions are caught
- UnmarkCommand.java : assert to ensure exceptions are caught
- TaskList.java : assert to ensure get and delete operation are of
  correct index
- Task.java : assert to ensure task description is not null

Beyond assertion, method names are optimized to comply with naming
standards.