Improve code quality
Currently most of the codes are complied with coding standard, but there
are a few more lines that is not of standard level.

With code quality below standard, it is hard for others to read and
understand the code, it is even hard for the author, to maintain and
modify the code.

Let's improve the code quality on following pieces of code:
- ExitCommand.java: adjust method name to start with verb
- Parser.java: adjust the if else statement to avoid deep nesting
- Ui.java: adjust method nmae to start with verb
Those modifications could help improve code readability and code
quality.

The detailed standard can be found here:
https://nus-cs2103-ay2122s2.github.io/website/schedule/week5/topics.html#readability