Refactor DateTimeParser
Refactor DateTimeParser

There are some usages of Magic Numbers used to reference indexes
of the input date-time string, and the main checkValidFormat method
is long (>30LOC).

Breaking the checkValidFormat method down into individual methods that
check for the separators and digits, time and date separately helps to
improve code quality.

Let's split it up so the quality of our code is maintained alongside
the rest of the repo.