Improve code quality
There are several areas of the code which consist of verbose statements,
as well as areas where the code was reused and redundant.

As a step towards better code quality, the redundant code was removed
and the code was refactored to coincide more with SLAP coding principles
for the sake of code readability.