improve quality of code
Currently, some parts of the code has long methods and deep nesting.

Quality of code needs to be improved to make it have better readability.

Changes made are as follows:
- In Parser, processMessage is divided into multiple short and simple methods
- In Storage, code is changed so that it is not in arrowhead style
- In TaskList, executeCommand is divided into multiple short and simple methods

By doing these, it will be easier for the readers to understand the code.

Other minor changes such as changing method names are made.