Improve Code quality based on rules stated in CS2103 website
Code quality is important because it helps improve readability of
code. This way, if other people were to be working on the same code,
they would have an easier time understanding the code. There is also
the benefit of the author himself understanding the code better.

For each of the files, rules based on the CS2103 website are
strictly adhered to to ensure readability of code.

Code quality is updated now as we have accumulated some technical
debt and it is about time to pay it.