Update gradle.yml file to exclude IO-redirection checks (A-CI)
Duke does not automate building and testing after each code change.

It is important for Duke to be able to do so as it allows the developer to identify bugs faster. It also makes the process of testing more seamless and efficient.

To facilitate automated testing and building, let's add automated workflows to test every push and pull request to the repository.

Setting up CI will allow easier shipment and deployment of code.