Refactor Task classes to support TaskList toFileString method
Storage Class requires all the classes under duke.Task package to be imported.

This is not ideal as it in addition to the places unecessary dependancy on the individual tasks classes, violating the Single Responsbility Principle of OOP.
It woud hence be more ideal for the storage class to only interact with the TaskList.

By providing a toFileString method, which coverts the tasks to a specificed file string format, TaskList can be improved to include a toFileString method
which converts its list of Tasks to a specified String format for file saving.

This ensures that the Storage class has minimal dependancy on the Task Classes, and only interacts with the TaskList class.
This potentially makes future debugging easier due to the reduced depedancies of classses.