Refactor UI class to be static
Methods under the UI class are not static, and would therefore require a instance to be created before its method could be accessed.

After reviewing the functionalities of the program, there is no need for there to be multiple instances of Ui.

By changing the methods of Ui to be static emulates the implementation of a singel Ui instance throughout the classes, so that any call refers to the same instance.

This removes the hassle of needing to create a Ui class just to acess its methods.