Package command and ui: use assertions
Assertions are not used to check whether the ui buffer is empty before a
new command is executed. Assertions are also not used to check whether
the program has exited.

Using assertions allows us to explicitly verify that the ui buffer is
empty at the state of program execution before new messages for a new
command is added. It also helps us ensure no response is received after
the program has exited.

Let's,
* Include a new function in Ui to check for empty buffer
* Use assertions and the new function to ensure the buffer is empty
before execution of any command
* Use assertions to ensure program does not receive response after it
has exited