Refactor Parser methods to be static
Similar to the UI class previously, methods under the Parser class are not static, and would therefore require a instance to be created before its method could be accessed.

The role of the Parser class is to simply provide the run and validate method, which does not gain any functionality from a instance being created.
Even when considering any future possible changes, there should not be a need where multiple Parser instances will be needed.

By changing the methods of Parser, namely validate and run to be static, it emulates the implementation of a singel Parser instance throughout the classes, ensuring consistency.

This removes the hassle of needing to create a Parser class just to acess its methods.