Add Exit command for GUI
Exit command has a behaviour (e.g. hasExit).

Closing the application without pressing
the 'X' button on the top left of the GUI allows the user
an alternative to close the application.

Implementing this behaviour gives the user an alternative to exit
the application with the Exit command. Furthermore, a delay of
1.5 seconds has been added after the exit command is typed by the
user. This means that after 1.5 seconds from entering the
Exit command, the application will automatically close.