Command response messages: encapsulate response messages inside a parent class CommandResponse
When a Command is executed by Spark, only a String object containing the response message is generated. This restricts metadata that can be specified about the response of the executed Command; such as whether it was produced as a result of a successful execution of the given Command, or if it was produced because of an exception produced when the Command was run.

Without this metadata, the user-interface cannot tailor the color of the chat box produced; ie it would not be possible to produce a red-colored chat box to indicate to the user that the input command could not be executed successfully.

To allow such additional information to be added, let's encapsulate the result of a Command into a class that not only contains the message to be shown to the user, and also other attributes of the result of the command's execution like whether it is an error, or if it is a warning that indicates some possibly undesirable impact resulting from its execution.

Doing so would provide additional visual feedback to the user about the result of the command's execution.