Fix Tasks not being updated properly when search function is used
Each TaskCard has an index which is used to access the Task it represents in the TaskList. The index assigned is the position of the TaskCard in the VBox.

When the search function is used, the position of the TaskCard does not match the index of the Task in the TaskList.

Let's perform filtering of Tasks in Window instead of TaskList.

This allows the original index of the filtered Tasks to be tracked and passed to each TaskCard.