Implement A-CodeQuality Current code is not clean enough and can be improved further.
In Parser.java and Storage.java, there are duplicate codes as well as
excessively long methods.

Let's create new methods that abstracts the creation of task, such
that Storage:loadData and Parser:parse will not be as long and will
be clearer. We can also create a new class that deals with checking
of date and time, which allows us to abstract the relevant methods
away from Parser and Storage.

In addition, we can also remove the Ui class and Storage:readData
since we have shifted to using GUI.