Refactor Main.java
There are three images used for the GUI. The user and innkeeper icons are set in MainWindow.java controller but background image is set in MainWindow.fxml by passing its url to the style prop of dialogContainer.

When the application is packaged into a .jar file and executed, the two icons appear fine but the background image does not load.

Let's,
* use getResourceAsStream() to retrieve background image in MainWindow.java controller
* set the background of dialogContainer in the initialize() method
* remove styles related to setting background of dialogBox in MainWindow.fxml

The relative URL to the background image becomes incorrect when the application is packaged into a .jar file. Moving the code for setting the background image from MainWindow.fxml to MainWindow.java allows for the reference to the background image to persist in the .jar file.