Add Java assertions
Add null-check assertions

There are many points in the code where objects declared are assumed
to be not null, that is, the error-handling code is presumed to work
correctly.

This is a dangerous assumption to make, and could result in hidden
bugs and errors arising from inconsistent code states.

Using a trust-but-verify approach, let's add assertions throughout
the application logic to ensure that the error-handling code works
correctly. Doing so flags events when the application state is
inconsistent, allowing bugs to be identified more easily.

Add default case false-assertions

There are several if-else statements that have default cases that are
ignored simply because they should never be reached. No code exists
within these empty cases.

This is also a dangerous assumption to make, as errors in code could
lead to unexpected behaviour.

Let's add false-assertions to these portions in the code to help
identify when these cases are triggered, which indicates that there are
bugs in the code logic.