Add assertions
It is important to check that the number of tasks in the list, after
adding one, is not 0. This verifies that the adding function to the
task list is working fine.

Furthermore, we need to check if the index we specified is based on
conventional array-based indexing and not the indexing used by the
user.

Lastly, we also need to check that tasks are stored properly so that
there will not be any parsing problems later on.