Add assert keywords to methods that contain switch statement, and other methods that does not include the else statement.
There are several methods that uses switch statament and assumes that
the default branch will not be invoked, or the if statment catches
the alternate assumption of the variable.

Adding the assert keywords will eliminate these
assumptions.

As a step towards eliminating assumptions of the program, let's include
the assert keyword to methods that assumes that a variable will be of
the correct value.