Improve code readability in command package
Change magic numbers and strings to constant variables

Magic numbers are used in command processing logic to translate from
a human readable format to a format that is used in the application's
internal state.

Magic numbers and strings can be difficult to decipher if the reader
has no idea what the code logic is trying to do, and negatively
impacts the readability of the codebase.

Let's extract the magic numbers and strings to named constants that
describe the function of the magic strings or constants. This makes
the code easier to read, and friendlier to readers not involved in
development.

Extract common functionality in CreateCommand to methods

The logic in CreateCommand involves 3 entirely separate flows, and are
implemented in 3 separate methods independent from one another.

However, some of the logic, particularly input validation, are
repeated across the 3 flows, resulting in almost identical code
portions.

Let's extract the common functionality into methods so there is only
a single implementation of the validation logic. This makes the logic
more maintainable and reusable.