Improve code readability in other packages
Break methods into smaller subroutines

The implementation in TaskListSerializer uses a single method to
perform all the serialization and deserialization required for
TaskList objects.

This results in a complex and deeply nested method, with many
subcomponents in the processing logic, which can be difficult to
understand when reading the code.

Let's break the long serialization methods into subroutines to make
the code more readable, and at the same time reduce the complexity
concentrated in a single part in code.