Add java stream function
TaskList method: use for loop to extract tasks by search term

TaskList find method creates a new ArrayList and use a for-loop to iterate through each element to find the task(s) that contains the search term.
 
Using a for-loop is more error-prone. Working with Streams enable us with higher level of abstraction. In addition, it generates more intuitive code.

Let's convert the method to implement Java's stream feature to streamline the search process.

Removes the need to perform multiple if-else checks by using a filter function and maps it a List collection.

Refer to this post on the advantages of using streams
https://www.capitalone.com/tech/software-engineering/java-streams-explained-simple-example/