Add assertions.
Assertions are used to ensure that the length of the TaskList increase
in length by 1 after each time the todo, deadline, even commands are
executed.

Assertions are used to define assumptions about the program state so
that the runtime can verify them. An assertion failure indicates a
possible bug in the code.