Add Java assertions
After adding a task to the task list of the Lister object,
it is expected that the size of the task list is more than zero.

The size of the task list should be checked after commands like
ToDo, Event and Deadline to ensure that the program is working as
intended.

Let's add assert statements to each of these commands to confirm
that the task list contains at least 1 item after their execution.