Implement Assertions into codebase
Codebase contains many try-catch statements as a form of extra protection
should the program not run as intended, enabling one to detect bugs.

Assertions are statement in the Java programming language that
enables one to test your assumptions about your program.

Writing assertions while programming is one of the
quickest and most effective ways to detect and correct bugs.

For a start, let's
- include assertions for switch statements with no default case.
- assert that a directory is created after calling the method mkdirs() in
the save method of Storge, removing the need for try-catch statements.