Implement  Assertions in the delete and add commands
The use of assertions helps to ensure that the add commands and delete commands is working as intended.
The size of task list should incremented by one when a task has been added.
The size of task list should be decremented by one when a task has been deleted.

The size of the task list is being compared with it's size after performing the add and delete commands
to ensure that the commands has been successful in executing their respective functions as intended.

Let's add assert statements to each addCommands (Todo, Deadline, Event) and deleteCommands to
ensure that the size of the task list contains one more or one less than the initial size of the task list
after performing the execution.